`infection/infection:0.16.3` requires PHP 7.3 or newer: killing off PHP 7.2 for good
There's no point in trying to keep compatibility with multiple infection versions: killing the
support for PHP 7.2 instead.

This also means that this particular change will not land in `4.3.0`, but in a later release
instead.

Locking `composer.json` platform to PHP 7.3.0 to be sure that we don't update dependencies
against something that is too new (most developers on the project are already on PHP 7.4.6+)