Removed need for manually checking for parsing errors on `composer.(json|lock)` and `installed.json`
This marginally reduces code size, and reduces `mixed` types a bit, but would benefit
tons from removing direct usage of `json_decode()`, which is an unsafe construct, when
the results of its calls go un-checked.

Ideally, we'd use `azjezz/psl`, but that library comes with added dependencies that we can't afford
here, for now.